Move functions to `dd.py`, add console preview option
